Handle the directory case more intelligently for linkbacks
When linking from dropbox to another path, we need to handle the initial
population case sanely -- that is, when the path in dropbox is still
a directory rather than a link, we need to take that content and move it into
the real home directory path. We were already doing this, but the current
implementation moved the directory out of the way and moved the dropbox path
into place, but this causes problems with directories whose ACLs are such that
they can't be removed (e.g. Library/Fonts). So, if the path in dropbox is
a directory, handle this specially and just copy its contents into the
existing directory.

Signed-off-by: Christopher Larson <kergoth@gmail.com>